rework synchronization
Even RLocks of the sync package are not reentrant.
Therefore we need a solution for calling alread synchronized methods
as part of an object implementation.

A basic implementation scheme is used to support this.
Every indicator consists of two objects a frontend object
acquiring the locks on its methods, and an implementation
object with the sme interface which finally implements the
functionality. This mechanism is also used to offer
some kind of protected methods, offered by the implementation
objects but not by the frontend object.

Every implementation object features a Seld object providing
access to the effective frontend and inmplementation object.

Additionally the blocks handling is streamlined.
All output by done now by a single go routine.